pipeline/workflows/prod_pipeline/requirements.txt
minio
duckdb
pandas
psycopg2-binary
apache-airflow-providers-cncf-kubernetes
pyspark
cassandra-driver
neo4j
requests
